refactor: 입력 수에 대한 검증 테스트 케이스 구현
입력 수에 대한 검증 테스트 케이스 구현
- 기존 Util에 두었던 입력 값에 대한 검증 기능 패키지 따로 구분
